create outputManager
