+++ cmd/single/main.go
@@ -80,6 +80,7 @@ func buildCommands() *cobra.Command {
 	envResolvers[env.Credential] = credResolver
 
 	inputManager := runner.NewInputManager(envResolvers, inputList, inputText, inputBool, inputPassword)
+	outputManager := runner.NewOutputManager(os.Stdout)
 	formulaSetup := runner.NewDefaultSingleSetup(ritchieHomeDir, http.DefaultClient)
 
 	defaultPreRunner := runner.NewDefaultPreRunner(formulaSetup)
@@ -87,7 +88,7 @@ func buildCommands() *cobra.Command {
 
 	postRunner := runner.NewPostRunner()
 
-	defaultRunner := runner.NewDefaultRunner(defaultPreRunner, postRunner, inputManager)
+	defaultRunner := runner.NewDefaultRunner(defaultPreRunner, postRunner, inputManager, outputManager)
 	dockerRunner := runner.NewDockerRunner(dockerPreRunner, postRunner, inputManager)
 
 	fileManager := stream.NewFileManager()
